+++ Kruglova/cw/text.txt
@@ -1,0 +1,6 @@
+-1
+qqqqq
+qqqqqqq
+qq
+
+https://repl.it/@kruglova9382/Kruglovacw#headers.h
