--- Kruglova/lab 5/text.txt
@@ -1,2 +1,0 @@
-0
-01100001001100010100110001101101110000001010
