--- Kruglova/lab 5/Инструкция по запуску .txt
@@ -1,8 +1,0 @@
-Тестировать код можно здесь:
-https://repl.it/@kruglova9382/Kruglovalab5#text.txt
-
-Для запуска введите в эмулированную консоль:
-./a.out <text.txt //для ввода из файла
-./a.out //для ввода с консоли
-
-
